+++ contracts/libraries/wrapppers/IdToAddressBiMapWrapper.sol
@@ -6,7 +6,7 @@ import "../IdToAddressBiMap.sol";
 contract IdToAddressBiMapWrapper {
     IdToAddressBiMap.Data private map;
 
-    function hasId(uint16 id) public view returns (bool) {
+    function hasId(uint24 id) public view returns (bool) {
         return IdToAddressBiMap.hasId(map, id);
     }
 
@@ -14,15 +14,15 @@ contract IdToAddressBiMapWrapper {
         return IdToAddressBiMap.hasAddress(map, addr);
     }
 
-    function getAddressAt(uint16 id) public view returns (address) {
+    function getAddressAt(uint24 id) public view returns (address) {
         return IdToAddressBiMap.getAddressAt(map, id);
     }
 
-    function getId(address addr) public view returns (uint16) {
+    function getId(address addr) public view returns (uint24) {
         return IdToAddressBiMap.getId(map, addr);
     }
 
-    function insert(uint16 id, address addr) public returns (bool) {
+    function insert(uint24 id, address addr) public returns (bool) {
         return IdToAddressBiMap.insert(map, id, addr);
     }
 }
